Server can now connect and receive transactions from the client
This is a POC. Currently the server can only really talk to one client
at a time. We need to rewrite it so that it can talk to N clients and
give a dedicated session to each one.

having the ability to send the output of the client's vm to the client
would be nice too.