Redirected vm output to client fifo
Now the console output of a client's vm is redirected to the client's
fifo. This make way more sence as provided the clients are remote they
wouldn't have access the the server's window.

Becides, isolating the clients would have been for nothing if the
clients were allowed to see each other vms.