Server: Fix broken pipe exception
Fix a broken pipe exception that would occur when freeing vms.
removeItem was trying to notify a client which was being deleted and
its pipe closed.

I changed the call to send() to use the MSG_NOSIGNAL flag to prevent
it from raising the SIGPIPE exception since clients disconnections are
being handled elsewhere.